# Only construct FnSymTab once, using `OnceCell`