--- a/partiql-logical-planner/Cargo.toml
+++ b/partiql-logical-planner/Cargo.toml
@@ -33,6 +33,7 @@ petgraph = "0.6.*"
 num = "0.4"
 fnv = "1"
 assert_matches = "1.5.*"
+once_cell = "1"
 
 [dev-dependencies]
 partiql-eval = { path = "../partiql-eval", version = "0.2.*" }