fixes #317 

for `cargo bench compile`:

before:
```
compile-1               time:   [26.914 µs 27.030 µs 27.158 µs]
                        change: [-1.4509% -0.1010% +1.2654%] (p = 0.88 > 0.05)
                        No change in performance detected.
Found 5 outliers among 100 measurements (5.00%)
  3 (3.00%) high mild
  2 (2.00%) high severe

compile-15              time:   [71.181 µs 71.398 µs 71.624 µs]
                        change: [+0.3953% +1.1365% +1.8644%] (p = 0.00 < 0.05)
                        Change within noise threshold.
Found 9 outliers among 100 measurements (9.00%)
  7 (7.00%) high mild
  2 (2.00%) high severe

compile-30              time:   [122.76 µs 123.42 µs 124.27 µs]
                        change: [-0.4880% +0.2882% +1.0441%] (p = 0.48 > 0.05)
                        No change in performance detected.
Found 9 outliers among 100 measurements (9.00%)
  4 (4.00%) high mild
  5 (5.00%) high severe
```


after:
```
compile-1               time:   [8.4798 µs 8.5462 µs 8.6514 µs]
                        change: [-68.689% -68.281% -67.733%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  1 (1.00%) low mild
  8 (8.00%) high mild
  2 (2.00%) high severe

compile-15              time:   [52.131 µs 52.400 µs 52.715 µs]
                        change: [-26.885% -26.134% -25.429%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 15 outliers among 100 measurements (15.00%)
  7 (7.00%) high mild
  8 (8.00%) high severe

compile-30              time:   [102.97 µs 103.48 µs 104.08 µs]
                        change: [-16.501% -15.725% -14.863%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 8 outliers among 100 measurements (8.00%)
  4 (4.00%) high mild
  4 (4.00%) high severe  
```

---

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.